requirements.txt
flask==2.3.3
google-generativeai==0.8.3
PyPDF2==3.0.1
pytesseract==0.3.13
python-docx==1.1.2
python-dotenv==1.0.1
werkzeug==3.0.4
Pillow==10.4.0